+++ package.json
@@ -49,6 +49,7 @@
     "classnames": "2.3.2",
     "clsx": "2.1.1",
     "dompurify": "3.1.0",
+    "dotenv": "16.4.7",
     "file-saver": "2.0.5",
     "final-form": "4.20.9",
     "focus-trap-react": "10.2.3",
@@ -63,6 +64,7 @@
     "moment": "2.30.1",
     "next": "13.4.5",
     "node-cron": "3.0.3",
+    "node-fetch": "3.3.2",
     "nodemailer": "6.9.14",
     "passport-orcid": "0.0.4",
     "postcss": "8.4.27",
